Update authentication.js
Was bothered by having html-editing type stuff in this file, feel like it belongs in the other js file. So register() and confirmRegistration() have calls to new functions i made in the other js file which will do appropriate html-editing stuff for log in menus. I am not super confident in this approach. Let me know if you think of a better way to do it. Not sure if there is a way for  the other js file (named lrc.js right now) to notice a successful sign up as an event, or something like that, as an alternative to this.